feature/applcation: add test case

+++ src/main/java/com/plongrotha/loanmanagement/service/impl/ApplicationServiceImpl.java
@@ -2,10 +2,14 @@ package com.plongrotha.loanmanagement.service.impl;
 
 import java.util.List;
 
+import org.springframework.cache.annotation.CacheEvict;
+import org.springframework.cache.annotation.Cacheable;
 import org.springframework.stereotype.Service;
 
+import com.plongrotha.loanmanagement.enums.LoanRefundStatus;
 import com.plongrotha.loanmanagement.exception.NotFoundException;
 import com.plongrotha.loanmanagement.model.Application;
+import com.plongrotha.loanmanagement.model.LoanApplication;
 import com.plongrotha.loanmanagement.repository.ApplicationRepository;
 import com.plongrotha.loanmanagement.repository.LoanApplicationRepository;
 import com.plongrotha.loanmanagement.service.ApplicationService;
@@ -31,23 +35,32 @@ public class ApplicationServiceImpl implements ApplicationService {
         return existingApplication;
     }
 
+    @Cacheable(value = "loan", key = "#applicationId")
     @Override
     public Application getApplicationById(Long applicationId) {
         return applicationRepository.findById(applicationId)
                 .orElseThrow(() -> new NotFoundException("Application with ID " + applicationId + " not found"));
     }
 
+    @CacheEvict(value = "loan", key = "#applicationId")
     @Override
     public void deleteApplication(Long applicationId) {
         boolean hasByAppicationId = loanApplicationRepository.existsByApplicationApplicationId(applicationId);
+        LoanApplication loanApplication = loanApplicationRepository.findByApplicationId(applicationId);
         if (hasByAppicationId) {
+        	if(loanApplication.getLoanRefundStatus() == LoanRefundStatus.IN_PROGRESS) {
+        		throw new IllegalArgumentException("this application is not yet refund completed");
+        	}
             loanApplicationRepository.deleteById(applicationId);
         }
+        
+        // when loan
         Application existingApplication = applicationRepository.findById(applicationId)
                 .orElseThrow(() -> new NotFoundException("Application with ID " + applicationId + " not found"));
         applicationRepository.delete(existingApplication);
     }
 
+    // list all the Application to the client
     @Override
     public List<Application> getAllApplications() {
         List<Application> applications = applicationRepository.findAll();
